# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Articles Service",
    "description": "API for indexing press articles and linking them to outlets, journalists, and topics",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3012"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "properties": {
              "formErrors": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "fieldErrors": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "required": [
              "formErrors",
              "fieldErrors"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "DiscoveredArticle": {
        "type": "object",
        "properties": {
          "articleId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the upserted article record"
          },
          "articleUrl": {
            "type": "string",
            "description": "URL of the discovered article",
            "example": "https://techcrunch.com/2025/03/15/ai-funding-roundup"
          },
          "title": {
            "type": "string",
            "nullable": true,
            "description": "Article title from OpenGraph or search result"
          },
          "snippet": {
            "type": "string",
            "nullable": true,
            "description": "Short excerpt from the article"
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExtractedAuthor"
            },
            "description": "Authors extracted via scraping + LLM analysis"
          },
          "publishedAt": {
            "type": "string",
            "nullable": true,
            "description": "Publication date extracted from article metadata",
            "example": "2025-03-15T10:00:00Z"
          }
        },
        "required": [
          "articleId",
          "articleUrl",
          "title",
          "snippet",
          "authors",
          "publishedAt"
        ]
      },
      "ExtractedAuthor": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "person",
              "organization"
            ],
            "description": "Whether this author is a person or an organization (news agency, editorial team, etc.)",
            "example": "person"
          },
          "firstName": {
            "type": "string",
            "description": "First name (empty string for organizations or single-name authors)",
            "example": "Sarah"
          },
          "lastName": {
            "type": "string",
            "description": "Last name, or full name for organizations",
            "example": "Perez"
          }
        },
        "required": [
          "type",
          "firstName",
          "lastName"
        ]
      },
      "DiscoverOutletArticlesBody": {
        "type": "object",
        "properties": {
          "outletDomain": {
            "type": "string",
            "minLength": 1,
            "description": "Domain of the outlet (e.g. techcrunch.com)",
            "example": "techcrunch.com"
          },
          "maxArticles": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20,
            "default": 10,
            "description": "Max articles to discover (default 10)"
          }
        },
        "required": [
          "outletDomain"
        ]
      },
      "DiscoverJournalistPublicationsBody": {
        "type": "object",
        "properties": {
          "journalistFirstName": {
            "type": "string",
            "minLength": 1,
            "description": "Journalist first name",
            "example": "Sarah"
          },
          "journalistLastName": {
            "type": "string",
            "minLength": 1,
            "description": "Journalist last name",
            "example": "Perez"
          },
          "outletDomain": {
            "type": "string",
            "minLength": 1,
            "description": "Outlet domain to scope the Google News search via site: filter (e.g. 'techcrunch.com')",
            "example": "techcrunch.com"
          },
          "maxResults": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20,
            "default": 10,
            "description": "Max publications to find (default 10, max 20)"
          }
        },
        "required": [
          "journalistFirstName",
          "journalistLastName",
          "outletDomain"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/discover/outlet-articles": {
      "post": {
        "operationId": "discoverOutletArticles",
        "summary": "Discover recent articles from an outlet via Google News + scraping, and create scoped discoveries",
        "description": "Pipeline endpoint: (1) searches Google News for recent articles from the given outlet domain, (2) scrapes each article URL to extract authors and publication dates via LLM, (3) upserts the articles in the database, and (4) creates discovery records scoped to the campaign (x-brand-id, x-campaign-id headers required). Returns the discovered articles with extracted author details.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Internal org UUID"
            },
            "required": true,
            "description": "Internal org UUID",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Internal user UUID"
            },
            "required": true,
            "description": "Internal user UUID",
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Run UUID from runs-service"
            },
            "required": true,
            "description": "Run UUID from runs-service",
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug for tracking"
            },
            "required": false,
            "description": "Workflow slug for tracking",
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug for tracking/filtering"
            },
            "required": false,
            "description": "Feature slug for tracking/filtering",
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated brand UUIDs for scoping",
              "example": "e0000000-0000-4000-8000-000000000001,e0000000-0000-4000-8000-000000000002"
            },
            "required": false,
            "description": "Comma-separated brand UUIDs for scoping",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Campaign UUID for inter-service propagation"
            },
            "required": false,
            "description": "Campaign UUID for inter-service propagation",
            "name": "x-campaign-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoverOutletArticlesBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Discovered articles with extracted authors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "articles": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DiscoveredArticle"
                      }
                    }
                  },
                  "required": [
                    "articles"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/discover/journalist-publications": {
      "post": {
        "operationId": "discoverJournalistPublications",
        "summary": "Discover recent publications by a journalist and create scoped discoveries",
        "description": "Pipeline endpoint: (1) searches Google News for recent articles by the given journalist (by name), (2) scrapes each article URL to extract authors and publication dates via LLM, (3) upserts the articles in the database, and (4) creates discovery records scoped to the campaign and journalist (x-brand-id, x-campaign-id headers required). Ideal for enriching pitch generation with a journalist's recent work.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Internal org UUID"
            },
            "required": true,
            "description": "Internal org UUID",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Internal user UUID"
            },
            "required": true,
            "description": "Internal user UUID",
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Run UUID from runs-service"
            },
            "required": true,
            "description": "Run UUID from runs-service",
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug for tracking"
            },
            "required": false,
            "description": "Workflow slug for tracking",
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug for tracking/filtering"
            },
            "required": false,
            "description": "Feature slug for tracking/filtering",
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated brand UUIDs for scoping",
              "example": "e0000000-0000-4000-8000-000000000001,e0000000-0000-4000-8000-000000000002"
            },
            "required": false,
            "description": "Comma-separated brand UUIDs for scoping",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Campaign UUID for inter-service propagation"
            },
            "required": false,
            "description": "Campaign UUID for inter-service propagation",
            "name": "x-campaign-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoverJournalistPublicationsBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Journalist publications with extracted authors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "articles": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DiscoveredArticle"
                      }
                    }
                  },
                  "required": [
                    "articles"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}